Update Subscribers Controller
Fixing all function

+++ src/controllers/subscribers.controller.js
@@ -4,7 +4,11 @@ const Subscribers = db.subscribers;
 exports.findAll = (req, res) => {
   Subscribers.find()
     .then((result) => {
-      res.send(result);
+      res.send({
+        message: "Subscribers successfully fetched.",
+        timestamp: new Date().toString(),
+        data: result,
+      });
     })
     .catch((err) => {
       res.status(500).send({
@@ -14,22 +18,45 @@ exports.findAll = (req, res) => {
 };
 
 exports.create = (req, res) => {
-  const subscribers = new Subscribers({
-    email: req.body.email,
-    startDate: new Date(),
-    endDate: "",
-    status: "active",
-  });
+  const email = req.body.email;
 
-  subscribers
-    .save()
+  if (!email) {
+    return res.status(400).send({
+      message: "Email is required.",
+    });
+  }
+
+  Subscribers.findOne({ email })
     .then((result) => {
-      res.status(200).send({
-        message: "Playlist successfully added.",
+      if (result) {
+        return res.status(400).send({
+          message: "Email already exists.",
+        });
+      }
+
+      const subscribers = new Subscribers({
+        email: email,
+        startDate: new Date().toString(),
+        status: "active",
       });
+
+      subscribers
+        .save()
+        .then((result) => {
+          res.status(200).send({
+            message: "Playlist successfully added.",
+            timestamp: new Date().toString(),
+            data: result,
+          });
+        })
+        .catch((err) => {
+          res.status(500).send({
+            message: err.message || "Some error while creating Subscribers.",
+          });
+        });
     })
     .catch((err) => {
-      res.status(409).send({
+      return res.status(500).send({
         message: err.message || "Some error while creating Subscribers.",
       });
     });
@@ -38,19 +65,42 @@ exports.create = (req, res) => {
 exports.findOne = (req, res) => {
   const id = req.params.id;
 
+  if (!id) {
+    return res.status(400).send({
+      message: "Subscriber ID is required.",
+    });
+  }
+
   Subscribers.findById(id)
     .then((result) => {
-      res.send(result);
+      if (!result) {
+        res.status(404).send({
+          message: "Subscriber not found",
+        });
+      }
+
+      res.send({
+        message: "Subscriber was found.",
+        timestamp: new Date().toString(),
+        data: result,
+      });
     })
     .catch((err) => {
       res.status(500).send({
         message: err.message || "Some error while showing Subscribers.",
+        timestamp: new Date().toString(),
       });
     });
 };
 
 exports.delete = (req, res) => {
   const id = req.params.id;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "Subscriber ID is required.",
+    });
+  }
 
   Subscribers.findByIdAndRemove(id)
     .then((result) => {
@@ -62,11 +112,42 @@ exports.delete = (req, res) => {
 
       res.send({
         message: "Playlist was deleted",
+        timestamp: new Date().toString(),
       });
     })
     .catch((err) => {
-      res.status(409).send({
+      res.status(500).send({
         message: err.message || "Some error while delete playlist.",
+      });
+    });
+};
+
+exports.update = (req, res) => {
+  const id = req.params.id;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "Subscriber ID is required.",
+    });
+  }
+
+  Subscribers.findByIdAndUpdate(id, req.body)
+    .then((result) => {
+      if (!result) {
+        res.status(404).send({
+          message: "Subscriber not found",
+        });
+      }
+
+      res.send({
+        message: "Subscriber was updated.",
+        timestamp: new Date().toString(),
+        data: result,
+      });
+    })
+    .catch((err) => {
+      res.status(500).send({
+        message: err.message || "Some error while updating Subscribers.",
       });
     });
 };
